Added the code to the app
This was a large project for CSSE483, Andriod development. It was a pair
project with Coco Liu which allowed students to make appointments with a
tutor, as well as fill out reviews of the tutoring session. It stored
the data in a firebase and had both a website and IOS system that
plugged into the firebase as well, however I was not involved in
implementing these.